[CHANGE] Split VoIP.py into VoIP.phone, VoIP.call and VoIP.error. [CHANGE] Changed all imports to reflect new file names. [FIX] Fixed missing PhoneStatus enum in __all__.
