+++ pyVoIP/VoIP/error.py
@@ -1,0 +1,17 @@
+__all__ = [
+    "InvalidRangeError",
+    "InvalidStateError",
+    "NoPortsAvailableError",
+]
+
+
+class InvalidRangeError(Exception):
+    pass
+
+
+class InvalidStateError(Exception):
+    pass
+
+
+class NoPortsAvailableError(Exception):
+    pass
